# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fork & Find Restaurant Discovery API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /restaurants:
    get:
      operationId: listRestaurants
      tags: [restaurants]
      summary: List restaurants with optional filters
      parameters:
        - name: city
          in: query
          schema:
            type: string
        - name: district
          in: query
          schema:
            type: string
        - name: cuisine
          in: query
          schema:
            type: string
        - name: bestFor
          in: query
          schema:
            type: string
        - name: evidenceLevel
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of restaurants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Restaurant"
    post:
      operationId: createRestaurant
      tags: [restaurants]
      summary: Create a new restaurant (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantInput"
      responses:
        "201":
          description: Created restaurant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
...
  /restaurants/{id}:
    get:
      operationId: getRestaurant
      tags: [restaurants]
      summary: Get restaurant by ID with dishes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restaurant with dishes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantWithDishes"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateRestaurant
      tags: [restaurants]
      summary: Update a restaurant (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantInput"
      responses:
        "200":
          description: Updated restaurant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteRestaurant
      tags: [restaurants]
      summary: Delete a restaurant (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Dishes ───────────────────────────────────────────────────────────────────
...
  /shortlist/{restaurantId}:
    post:
      operationId: addToShortlist
      tags: [shortlist]
      summary: Add restaurant to shortlist
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Added to shortlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
    delete:
      operationId: removeFromShortlist
      tags: [shortlist]
      summary: Remove restaurant from shortlist
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed from shortlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Admin ─────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Restaurant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: string
        district:
          type: ["string", "null"]
        cuisine:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        googleMapsUrl:
          type: ["string", "null"]
        instagramUrl:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        menuSourceUrl:
          type: ["string", "null"]
        ratingSourceNotes:
          type: ["string", "null"]
        openingHoursNotes:
          type: ["string", "null"]
        reviewConsensusSummary:
          type: ["string", "null"]
        evidenceLevel:
          type: ["string", "null"]
          description: "strong / moderate / weak"
        bestFor:
          type: ["string", "null"]
          description: "Comma-separated: date, business, family, fine dining, casual"
        strengths:
          type: ["string", "null"]
        weaknesses:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - city
        - cuisine
        - isFeatured
        - createdAt
        - updatedAt
    RestaurantInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        city:
          type: string
          minLength: 1
        district:
          type: string
        cuisine:
          type: string
          minLength: 1
        latitude:
          type: number
        longitude:
          type: number
        googleMapsUrl:
          type: string
        instagramUrl:
          type: string
        websiteUrl:
          type: string
        menuSourceUrl:
          type: string
        ratingSourceNotes:
          type: string
        openingHoursNotes:
          type: string
        reviewConsensusSummary:
          type: string
        evidenceLevel:
          type: string
        bestFor:
          type: string
        strengths:
          type: string
        weaknesses:
          type: string
        isFeatured:
          type: boolean
      required:
        - name
        - city
        - cuisine
    RestaurantWithDishes:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: string
        district:
          type: ["string", "null"]
        cuisine:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        googleMapsUrl:
          type: ["string", "null"]
        instagramUrl:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        menuSourceUrl:
          type: ["string", "null"]
        ratingSourceNotes:
          type: ["string", "null"]
        openingHoursNotes:
          type: ["string", "null"]
        reviewConsensusSummary:
          type: ["string", "null"]
        evidenceLevel:
          type: ["string", "null"]
        bestFor:
          type: ["string", "null"]
        strengths:
          type: ["string", "null"]
        weaknesses:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        dishes:
          type: array
          items:
            $ref: "#/components/schemas/Dish"
      required:
        - id
        - name
        - city
        - cuisine
        - isFeatured
        - createdAt
        - updatedAt
        - dishes

    # ── Dish ────────────────────────────────────────────────────────────────────
    Dish:
      type: object
      properties:
        id:
          type: integer
        restaurantId:
          type: integer
        name:
          type: string
        category:
          type: string
          description: "starter / salad / sushi-raw / main / dessert"
        rawOrCooked:
          type: ["string", "null"]
          description: "raw / cooked"
        description:
          type: ["string", "null"]
        evidenceType:
          type: ["string", "null"]
          description: "signature / customer_favorite / critic_favorite / weak_evidence"
        evidenceLevel:
          type: ["string", "null"]
          description: "strong / moderate / weak"
        recommendationScore:
          type: ["integer", "null"]
          description: "1-10"
        dietTags:
          type: ["string", "null"]
          description: "Comma-separated: seafood, meat, chicken, vegetarian, spicy, light, indulgent"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - restaurantId
        - name
        - category
        - createdAt
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # ── Admin ────────────────────────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── OpenAI ────────────────────────────────────────────────────────────────────
...