# lib/api-zod/src/generated/types/restaurant.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fork & Find Restaurant Discovery API
 * OpenAPI spec version: 0.1.0
 */

export interface Restaurant {
  id: number;
  name: string;
  city: string;
  /** @nullable */
  district?: string | null;
  cuisine: string;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  googleMapsUrl?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  menuSourceUrl?: string | null;
  /** @nullable */
  ratingSourceNotes?: string | null;
  /** @nullable */
  openingHoursNotes?: string | null;
  /** @nullable */
  reviewConsensusSummary?: string | null;
  /**
     * strong / moderate / weak
     * @nullable
     */
  evidenceLevel?: string | null;
  /**
     * Comma-separated: date, business, family, fine dining, casual
     * @nullable
     */
  bestFor?: string | null;
  /** @nullable */
  strengths?: string | null;
  /** @nullable */
  weaknesses?: string | null;
  isFeatured: boolean;
  createdAt: Date;
  updatedAt: Date;
}